requirements.txt
discord
pygsheets
python-dotenv
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger-proto-grpc
opentelemetry-exporter-jaeger-thrift
